+++ script/rockgen.cpp
@@ -31,7 +31,7 @@ float RockGen::Fractal( int x, int y, float octave0, float octave1, float octave
 	float nx1 = OCTAVE_1 * (float)x * INV;
 	float ny1 = OCTAVE_1 * (float)y * INV;
 
-	float n = 1.0f - fabs( noise0->Noise( nx0, ny0, 0 ) - noise1->Noise( nx1, ny1, 0 ) );
+	float n = 1.0f - fabsf( noise0->Noise( nx0, ny0, 0.0f ) - noise1->Noise( nx1, ny1, 0.0f ) );
 	n = Clamp( n, -1.0f, 1.0f );
 	return n;
 }
